+++ php-backend/dbactions/.htaccess
@@ -1,0 +1,1 @@
+DENY FROM ALL
