+++ svelte-frontend/package.json
@@ -20,7 +20,9 @@
     "svelte": "^3.55.0"
   },
   "dependencies": {
+    "axios": "^1.6.3",
     "pg": "^8.11.3",
-    "sirv-cli": "^2.0.0"
+    "sirv-cli": "^2.0.0",
+    "svelte-spa-router": "^4.0.0"
   }
 }
