+++ svelte-frontend/src/routes/HomePage.svelte
@@ -1,0 +1,1 @@
+HOMEPAGE KOO
